+++ common/src/main/java/net/spell_engine/compat/CompatFeatures.java
@@ -7,6 +7,7 @@ public class CompatFeatures {
         ContainerCompat.init();
         CombatRollCompat.init();
         MeleeCompat.init();
+        CriticalStrikeCompat.init();
         FTBTeamsCompat.init();
     }
 }